+++ CHANGE_LOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- GH-17 Update to more recent bouncycastle library iterations, and switch to PEMParser from PEMReader
 
 ## [0.3.0]
 ### Added